pyproject.toml
[project]
name = "formatapi"
version = "0.1.0"
description = "AI API 配置格式化工具 - 智能解析和格式化AI服务配置"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # GUI框架
    "PyQt6>=6.6.1",
    # 图片处理
    "Pillow>=10.2.0",
    # 格式化输出
    "PyYAML>=6.0.1",
    "toml>=0.10.2",
    # 数据验证
    "pydantic>=2.5.3",
    # 工具库
    "pyperclip>=1.8.2",
    "requests>=2.31.0",
    # AI模型调用
    "openai>=1.10.0",
    "pyobjc-framework-vision>=12.1",
]

[project.optional-dependencies]
# macOS系统OCR支持（可选）
macos-ocr = [
    "pyobjc-framework-Vision>=10.1",
    "pyobjc-framework-Quartz>=10.1",
]

# 开发依赖
dev = [
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ui", "services", "utils"]
include = ["main.py", "config.py"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ['py310']
